backend/requirements.txt
typing_extensions==4.3.0
fastapi==0.85.0
fastapi>=0.80.0,<=0.85.0
uvicorn==0.18.3
python-multipart==0.0.5
pillow==9.2.0
requests==2.28.1
mysql-connector-python
python-dotenv